Add createHoppings and constructHamiltonians
This commit includes the source code for creating the Hoppings graph and creating
the Hamiltonian graph. The Hoppings requires a helper file named Materials that
determines how electrons move depending on the material. Construct Hamiltonian
builds on the inputed values generated by createHoppings to create the Hamiltonian.

The source code is paired with tests that test the functionality. we have passed in a
supercell that represents one unit cell, and check to see if all directions are present
when we create the Hoppings Graph. For the Hamiltonian, it outputs a 2x2 sparse graph
because there's spin included in the creation of it.